Added unit test for email validation, and centered submit button on sign up/sign in screens
