+++ package.json
@@ -7,7 +7,12 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "eject": "expo eject"
+    "eject": "expo eject",
+    "test": "jest"
+  },
+  "jest": {
+    "preset": "react-native",
+    "setupFiles": ["./src/setupTests.js"]
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.16.1",
@@ -28,6 +33,8 @@
     "expo-media-library": "~14.0.0",
     "expo-sharing": "~10.1.0",
     "expo-status-bar": "~1.2.0",
+    "jest": "^26.6.3",
+    "jest-expo": "^44.0.1",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-hook-form": "^7.26.0",
@@ -54,7 +61,12 @@
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@iconify/react": "^3.2.0",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/react-native": "^9.1.0",
     "@types/react-native": "~0.64.12",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
+    "enzyme": "^3.11.0",
+    "react-test-renderer": "17",
     "typescript": "~4.3.5"
   },
   "private": true
